splites tests as Library Modules
