+++ FileConverterLib.Tests/UnitTestsPdf.cs
@@ -11,82 +11,8 @@ using System.Net;
 namespace FileConverterLib.Tests
 {
     [TestClass]
-    public class UnitTests
+    public class UnitTestsPdf: AbstractUnitTests
     {
-        private static string pathTestfiles = Path.GetFullPath(Path.Combine("..", "..", "..", "..", "FileConverterLib.Tests", "testfiles"));
-        private static string pathResult = Path.GetFullPath(Path.Combine("..", "..", "..", "..", "FileConverterLib.Tests", "resultfiles"));
-
-        [ClassInitialize]
-        public static void BeforeTests(TestContext _context)
-        {
-            //создаём папку в которой будут хранится все файлы для тестов
-            if (!Directory.Exists(pathResult))
-                Directory.CreateDirectory(pathResult);
-
-            // удаляем все тестовые файлы
-            foreach (string f in Directory.GetFiles(pathResult))
-                File.Delete(f);
-
-            //удаляем все тестовые папки
-            foreach(string d in Directory.GetDirectories(pathResult))
-                Directory.Delete(d,true);
-        }
-
-        #region Test_PngToJpg
-
-        [TestMethod]
-        public void Test_PngToJpg_2params()
-        {
-            string filename = "test_png";
-            ImageConverter.PngFileToJpgFile(Path.Combine(pathTestfiles, filename), Path.Combine(pathResult, filename));
-
-            //проверяем сущестсвует ли файл с таким путём
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult,$"{filename}.jpg")));
-
-        }
-
-        [TestMethod]
-        public void Test_PngToJpg_1param()
-        {
-            string filename = "test_png";
-            File.Copy(Path.Combine(pathTestfiles, $"{filename}.png"), Path.Combine(pathResult, $"{filename}.png"));
-            ImageConverter.PngFileToJpgFile(Path.Combine(pathResult, filename));
-
-            //проверяем сущестсвует ли файл с таким путём
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult,$"{filename}.jpg")));
-
-        }
-
-        #endregion
-
-        #region Test_JpgToPng
-        [TestMethod]
-        public void Test_JpgToPng_2params()
-        {
-            string filename = "test_jpg";
-            ImageConverter.JpgFileToPngFile(Path.Combine(pathTestfiles, filename), Path.Combine(pathResult, filename));
-
-            //проверяем сущестсвует ли файл с таким путём
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.png")));
-
-        }
-
-        [TestMethod]
-        public void Test_JpgToPng_1param()
-        {
-            string filename = "test_jpg";
-
-            //копируем файл из папки testfiles в папку results, чтобы создать преобразованный файл в папке с исходным и с тем же названием
-            File.Copy(Path.Combine(pathTestfiles, $"{filename}.jpg"), Path.Combine(pathResult, $"{filename}.jpg"));
-            ImageConverter.JpgFileToPngFile(Path.Combine(pathResult, filename));
-
-            //проверяем сущестсвует ли файл с таким путём
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.png")));
-
-        }
-
-        #endregion
-
         #region Test_PdfToJpg
         [TestMethod]
         public void Test_PdfToJpgs_folder_2param()
@@ -348,150 +274,6 @@ namespace FileConverterLib.Tests
                     Assert.AreEqual((document.PageCount - splitFrom) + 1, doc2.PageCount);
                 }
             }
-        }
-        #endregion
-
-        #region Test_WordToPdf_MSOffice
-        [TestMethod]
-        public void Test_WordToPdf_MSOffice_2params()
-        {
-            string filename = "test_word";
-            MSOfficeConverter.DocxFileToPdfFile(Path.Combine(pathTestfiles, filename), Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.pdf")));
-
-        }
-
-        [TestMethod]
-        public void Test_WordToPdf_MSOffice_1param()
-        {
-            string filename = "test_word";
-            File.Copy(Path.Combine(pathTestfiles, $"{filename}.docx"), Path.Combine(pathResult, $"{filename}.docx"));
-            
-            MSOfficeConverter.DocxFileToPdfFile(Path.Combine(pathResult, filename));
-            
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.pdf")));
-
-        }
-        #endregion
-
-        #region Test_PptxToPdf_MSOffice
-        [TestMethod]
-        public void Test_PptxToPdf_MSOffice_2params()
-        {
-            string filename = "test_powerpoint";
-            MSOfficeConverter.PptxFileToPdfFile(Path.Combine(pathTestfiles, filename), Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.pdf")));
-
-        }
-
-        [TestMethod]
-        public void Test_PptxToPdf_MSOffice_1param()
-        {
-            string filename = "test_powerpoint";
-            File.Copy(Path.Combine(pathTestfiles, $"{filename}.pptx"), Path.Combine(pathResult, $"{filename}.pptx"));
-
-            MSOfficeConverter.PptxFileToPdfFile(Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.pdf")));
-
-        }
-        #endregion
-
-        #region Test_PdfToWord_MSOffice
-        [TestMethod]
-        public void Test_PdfToWord_MSOffice_2params()
-        {
-            string filename = "test_pdf_2";
-            MSOfficeConverter.PdfFileToDocxFile(Path.Combine(pathTestfiles, filename), Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.docx")));
-
-        }
-
-        [TestMethod]
-        public void Test_PdfToWord_MSOficce_1param()
-        {
-            string filename = "test_pdf_2";
-            File.Copy(Path.Combine(pathTestfiles, $"{filename}.pdf"), Path.Combine(pathResult, $"{filename}.pdf"));
-
-            MSOfficeConverter.PdfFileToDocxFile(Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.docx")));
-
-        }
-        #endregion
-
-        #region Test_PdfToWord_LibreOffice
-        [TestMethod]
-        public void Test_PdfToWord_LibreOffice_2params()
-        {
-            string filename = "test_pdf_1";
-            LibreOfficeConverter.PdfFileToDocxFile(Path.Combine(pathTestfiles, filename), Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.docx")));
-
-        }
-
-        [TestMethod]
-        public void Test_PdfToWord_LibreOffice_1param()
-        {
-            string filename = "test_pdf_1";
-            File.Copy(Path.Combine(pathTestfiles, $"{filename}.pdf"), Path.Combine(pathResult, $"{filename}.pdf"));
-
-            LibreOfficeConverter.PdfFileToDocxFile(Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.docx")));
-
-        }
-        #endregion
-
-        #region Test_WordToPdf_LibreOffice
-        [TestMethod]
-        public void Test_WordToPdf_LibreOffice_2params()
-        {
-            string filename = "test_word";
-            LibreOfficeConverter.DocxFileToPdfFile(Path.Combine(pathTestfiles, filename), Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.pdf")));
-
-        }
-
-        [TestMethod]
-        public void Test_WordToPdf_LibreOffice_1param()
-        {
-            string filename = "test_word";
-            File.Copy(Path.Combine(pathTestfiles, $"{filename}.docx"), Path.Combine(pathResult, $"{filename}.docx"));
-
-            LibreOfficeConverter.DocxFileToPdfFile(Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.pdf")));
-
-        }
-        #endregion
-
-        #region Test_PptxToPdf_MSOffice
-        [TestMethod]
-        public void Test_PptxToPdf_LibreOffice_2params()
-        {
-            string filename = "test_powerpoint";
-            LibreOfficeConverter.PptxFileToPdfFile(Path.Combine(pathTestfiles, filename), Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.pdf")));
-
-        }
-
-        [TestMethod]
-        public void Test_PptxToPdf_LibreOffice_1param()
-        {
-            string filename = "test_powerpoint";
-            File.Copy(Path.Combine(pathTestfiles, $"{filename}.pptx"), Path.Combine(pathResult, $"{filename}.pptx"));
-
-            LibreOfficeConverter.PptxFileToPdfFile(Path.Combine(pathResult, filename));
-
-            Assert.IsTrue(File.Exists(Path.Combine(pathResult, $"{filename}.pdf")));
-
         }
         #endregion
     }
